Refactor: remove all javascript style shorthand syntax.
The shorthand is handy, but they can be dangerous as they are block and not obj-c method.

If you chain the operation and one of the object is nil (such as parent in node.parent.children.query(@"//*")),
the app will crash in query() as the pointer is nil.

Rather than checking null hopelessly, i rather remove all these shorthands.