Adjustments to the Gemfile and Podfile.
I'm not out to totally overhaul the project -- mostly I want to get the project to build again. We need the latest cocoapods gem in order to support working with Xcode 7, so I basically pinned the other gems to whatever I was in the original Gemfile.lock, but went for the latest cocoapods gem. Thus we keep most everything else the same, just updating the one gem. Then there were Podfile issues with the JavaScriptCoreOpalAdditions, but if I got it directly from the repo instead of the normal spec, it worked; go figure. It's still the same desired version of the Additions.